PostPulse/requirements.txt
Jinja2
fastapi
uvicorn
google-generativeai
python-dotenv
